requirements.txt
PIL
django_thumbs